Rename repair.img to repair.zst and update it
Symlink to repair/boot/repair.zst from the boot/ directory of all hosts
where it should be installed in /boot for emergency purposes.

Add curl to the requirements passed to pkg system since it is no longer
included in base.

Mount /dev/pts as well as /dev on boot, and don't try to unmount the
rootfs on shutdown as this will always fail.