Update ap1300 to linux 6.7 and uboot 2024.01
Adjust the ipqess ethernet patch to reflect upstream dsa changes.

Pad the compressed initramfs used for tftp booting by 4k to work around
a bootloader bug which truncates the downloaded image. With this padding,
it is safe to return to zstd compression for the cpio archive, shrinking
the tftp boot image.